Multiply graphs on common scale v1

+++ lib/Reader/IntervalBuffer.cpp
@@ -5,6 +5,7 @@
 #include "IntervalBuffer.h"
 
 #include <iostream>
+#include <limits>
 
 #include "PseudoSyncPrmBuffer.h"
 
@@ -142,7 +143,7 @@ std::unique_ptr<Reader_ifs::Chunk> IntervalBuffer::getPoints(const Borders &bord
     auto step = buff_step / target_step;
 
     size_t begin_offset = start_pos_ + size_t(buff_step * double((cr_borders.begin - borders_.begin).time));
-    size_t first_end_offset = begin_offset + size_t(step * target_len);
+    size_t first_end_offset = begin_offset + size_t(step * double(target_len));
 
     size_t second_end_offset = 0;
 
@@ -158,6 +159,7 @@ std::unique_ptr<Reader_ifs::Chunk> IntervalBuffer::getPoints(const Borders &bord
 
     size_t pos = 0;
     size_t transition_index = begin_offset;
+
 
     for (size_t i = begin_offset; i < first_end_offset; i++) {
         Point &val = data_[i];
